fix: 데이터 꼬임에러 수정

+++ src/firebase.js
@@ -1,9 +1,6 @@
 import { initializeApp } from "firebase/app"
-import { deleteDoc, getFirestore } from "firebase/firestore"
-import { doc } from "firebase/firestore"
-import { setDoc } from "firebase/firestore"
-import { collection, getDocs } from "firebase/firestore";
-import {  updateDoc, getDoc, increment } from "firebase/firestore";
+import { doc, setDoc, collection, getDoc, getDocs }from "firebase/firestore"
+import {updateDoc, increment, deleteDoc, getFirestore } from "firebase/firestore"
 import firebaseConfig from '../firebaseConfig.js'
 initializeApp(firebaseConfig);
 
@@ -17,11 +14,12 @@ async function getTotalData() {
   // 키워드별 센텐스 빈 배열 생성
   for (let i in groupNameList) {
     const collections = await getDocs(collection(db, groupNameList[i]))
-    collections.forEach((document) => {
-      let keyword = document.id
+    let docs = collections.docs
+    for (let d=0; d< docs.length; d++ ){
+      let keyword = docs[d].id
       result[keyword] = {}
       result[keyword]["sentences"] = []
-    })
+    }
   }
 
   // 키워드별 sentence 삽입 
@@ -46,29 +44,31 @@ async function getTotalData() {
       }
     }
   }
-  console.log(result)
   return result
 }
 
-// 기존 데이터(키워드, 그룹) 삭제
+// // 기존 데이터(키워드, 그룹) 삭제
 async function deleteOriginalData() {
   let removeList = []
   for (let word of ["Groups", "Keywords"]) {
     const collections = await getDocs(collection(db, word))
     collections.forEach((document) => {
-      let word = document.id
-      removeList.push(word)
+      let removeWord = document.id
+      removeList.push(removeWord)
+      deleteDoc(doc(db, word, removeWord ))
     })
   }
-  
+
   for (let word of removeList) {
     const collections = await getDocs(collection(db, word))
     let docs = collections.docs
     for(let d=0; d<docs.length; d++) {
       let keyword = docs[d].id
-      deleteDoc(doc(db, word, keyword))
+      if (keyword != undefined)
+        deleteDoc(doc(db, word, keyword))
     }
   }
+
 }
 
 // 데이터 추가
@@ -77,10 +77,6 @@ async function setData(dataCollection) {
   let groups = []
   // 기존 데이터 삭제
   await deleteOriginalData()
-  for (let i in dataCollection) {
-    groups.push(dataCollection[i]["group"])
-  }
-  await setGroups(groups)
   
   // 그룹별로 저장
   for (let i in dataCollection) {
@@ -91,6 +87,7 @@ async function setData(dataCollection) {
 
     // 키워드 따로 저장
     keywords.push(keyword)
+    groups.push(group)
 
     data["totalCount"] = dataCollection[i]["totalCount"]
 
@@ -122,6 +119,12 @@ async function setData(dataCollection) {
   for (let i in keywords) {
     await setDoc(doc(db, "Keywords", keywords[i]), {"keywords": keywords[i]})
   }
+
+  // 그룹만 저장
+  groups = Array.from(new Set(groups))
+  for (let i in groups) {
+    await setDoc(doc(db, "Groups", groups[i]), {"groupName": groups[i]})
+  }
 }
 
 // 키워드별 토탈카운트
@@ -143,19 +146,22 @@ async function getDataByGroups() {
   // 키워드당 빈 배열을 만들어주기
   for (let i in groupNameList) {
     const collections = await getDocs(collection(db, groupNameList[i]))
-    collections.forEach((document) => {
-      result[document.id] = {}
-    })
+    let docs = collections.docs
+    for (let d=0; d< docs.length; d++ ){
+      result[docs[d].id] = {}
+    }
   }
 
   // 그룹 순회하며 하나씩 가져오기
   for (let i in groupNameList) {
     const collections = await getDocs(collection(db, groupNameList[i]))
     let group = groupNameList[i]
-    collections.forEach((document) => {
-      let keyword = document.id
-      result[keyword][group] = document.data().totalCount
-    })
+    let docs = collections.docs
+    for (let d=0; d< docs.length; d++ ){
+      let keyword = docs[d].id
+      result[keyword][group] = docs[d].data().totalCount
+    }
+
   }
   return result
 }
@@ -173,13 +179,6 @@ async function getPassword () {
 
 // 그룹 데이터 저장
 async function setGroups (groups) {
-  // 원래 있던 데이터 삭제
-  const collections = await getDocs(collection(db, "Groups"))
-  collections.forEach((document) => {
-    deleteDoc(doc(db, "Groups", document.data().groupName))
-  })
-
-  // 들어온 그룹으로 설정
   for (let group of groups) {
     setDoc(doc(db, "Groups", group), {"groupName": group})
   }
@@ -215,6 +214,7 @@ async function updateCount(group, keyword, sentenceId) {
     }
 
     updateDoc(doc(db, group, keyword), changeContent)
+  
 }
 
 async function getCurrentData() {
